+++ src/main/java/ru/rerumu/lists/controller/book/view/in/BookUpdateView.java
@@ -26,30 +26,6 @@ public class BookUpdateView {
     private final List<Long> tagIds;
     private final List<ReadingRecordView> readingRecords;
 
-    @Getter
-    public static class ReadingRecordView {
-        private final Long readingRecordId;
-        private final Long statusId;
-        private final LocalDateTime startDate;
-        private final LocalDateTime endDate;
-        private final Long lastChapter;
-
-        @JsonCreator
-        public ReadingRecordView(
-                Long readingRecordId,
-                Long statusId,
-                @NonNull LocalDateTime startDate,
-                LocalDateTime endDate,
-                Long lastChapter
-        ) {
-            this.readingRecordId = readingRecordId;
-            this.statusId = statusId;
-            this.startDate = startDate;
-            this.endDate = endDate;
-            this.lastChapter = lastChapter;
-        }
-    }
-
     @JsonCreator
     public BookUpdateView(
             Long readListId,
@@ -77,5 +53,30 @@ public class BookUpdateView {
         this.URL = URL;
         this.tagIds = tagIds;
         this.readingRecords = readingRecords;
+    }
+
+    @ToString
+    @Getter
+    public static class ReadingRecordView {
+        private final Long readingRecordId;
+        private final Long statusId;
+        private final LocalDateTime startDate;
+        private final LocalDateTime endDate;
+        private final Long lastChapter;
+
+        @JsonCreator
+        public ReadingRecordView(
+                Long readingRecordId,
+                Long statusId,
+                @NonNull LocalDateTime startDate,
+                LocalDateTime endDate,
+                Long lastChapter
+        ) {
+            this.readingRecordId = readingRecordId;
+            this.statusId = statusId;
+            this.startDate = startDate;
+            this.endDate = endDate;
+            this.lastChapter = lastChapter;
+        }
     }
 }