+++ src/main/java/ru/rerumu/lists/dao/base/PersistenceProxy.java
@@ -2,7 +2,9 @@ package ru.rerumu.lists.dao.base;
 
 import lombok.Getter;
 import lombok.Setter;
+import lombok.ToString;
 
+@ToString(doNotUseGetters = true)
 public abstract class PersistenceProxy<T> {
 
     @Getter